Change ints and chars to uint32_t, int32_t and uint8_t.
Bytecode, strings and raw bytes still use char and unsigned char. This
isn't an issue since char has the same width for all platforms.

+++ tools.py
@@ -1,4 +1,4 @@
-import os, subprocess, sys, shutil
+import os, os.path, subprocess, sys, shutil
 
 USAGE = 'usage: python3 tools.py [build | buildDebug | clean | test | valgrind | help]'
 
@@ -156,7 +156,8 @@ def build(debugging=False):
     return not failed
 
 def clean():
-    shutil.rmtree('build')
+    if os.path.exists('build'):
+        shutil.rmtree('build')
 
 def test():
     if build(True):
